requirements.txt
activecampaign-python==1.0.3
certifi==2021.5.30
charset-normalizer==2.0.6
idna==3.2
mysql-connector-python
python-dotenv==0.19.0
requests==2.26.0
urllib3==1.26.7
